+++ app/CoreIntegrationApi/RestApi/RestRequestDataPrepper.php
@@ -9,10 +9,15 @@ class RestRequestDataPrepper extends RequestDataPrepper
     public function prep()
     {
         // possess rest request
-    }
+        // ! prep this data ************************************************
 
-    public function getPreppedData()
-    {
-        // get rest request
+        // $this->class = $request['class'];
+        //     $this->endpoint = $request['endpoint'];
+        //     $this->methodCalls = $request['methodCalls'] ?? [];
+        //     $this->includes = $request['includes'] ?? [];
+        //     $this->perPageParameter = $request['perPageParameter'] ?? 30;
+        //     $this->orderByParameter = $request['orderByParameter'] ?? [];
+        //     $this->selectParameter = $request['selectParameter'] ?? [];
+        //     $this->otherParameter = $request['otherParameter'] ?? [];
     }
 }
