+++ app/CoreIntegrationApi/RestApi/RestRequestProcessor.php
@@ -5,6 +5,6 @@ namespace App\CoreIntegrationApi\RestApi;
 use App\CoreIntegrationApi\RequestProcessor;
 
 class RestRequestProcessor extends RequestProcessor {
-    // uses serves provider Located ...
-    // loads __construct(RestRequestValidator $requestValidator, RestQueryResolver $queryResolver, RestResponseBuilder $responseBuilder)
+    // uses serves provider Located app\Providers\RestRequestProcessorProvider.php
+    // As of 1/1/22 loads __construct(RestRequestValidator $requestValidator, RestQueryResolver $queryResolver, RestResponseBuilder $responseBuilder)
 }
