+++ config/app.php
@@ -178,7 +178,9 @@ return [
         /*
          * Custom Service Providers...
          */
+        App\Providers\CILProvider::class,
         App\Providers\RestRequestProcessorProvider::class,
+        App\Providers\ContextRequestProcessorProvider::class,
     ],
 
     /*
